Move ReportCron class, update docs
Moved ReportCron class to reportengine.py, removed cronparse.py
Minor changes to customization guide
